+++ gpdb-20160503/src/Models/DescribeDBVersionRequest.php
@@ -1,0 +1,62 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Gpdb\V20160503\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class DescribeDBVersionRequest extends Model
+{
+    /**
+     * @var string
+     */
+    public $DBInstanceId;
+
+    /**
+     * @var int
+     */
+    public $ownerId;
+    protected $_name = [
+        'DBInstanceId' => 'DBInstanceId',
+        'ownerId' => 'OwnerId',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->DBInstanceId) {
+            $res['DBInstanceId'] = $this->DBInstanceId;
+        }
+
+        if (null !== $this->ownerId) {
+            $res['OwnerId'] = $this->ownerId;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['DBInstanceId'])) {
+            $model->DBInstanceId = $map['DBInstanceId'];
+        }
+
+        if (isset($map['OwnerId'])) {
+            $model->ownerId = $map['OwnerId'];
+        }
+
+        return $model;
+    }
+}