+++ gpdb-20160503/src/Models/DescribeDBVersionResponseBody.php
@@ -1,0 +1,62 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Gpdb\V20160503\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class DescribeDBVersionResponseBody extends Model
+{
+    /**
+     * @var string
+     */
+    public $requestId;
+
+    /**
+     * @var string
+     */
+    public $versionSuggestion;
+    protected $_name = [
+        'requestId' => 'RequestId',
+        'versionSuggestion' => 'VersionSuggestion',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+
+        if (null !== $this->versionSuggestion) {
+            $res['VersionSuggestion'] = $this->versionSuggestion;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+
+        if (isset($map['VersionSuggestion'])) {
+            $model->versionSuggestion = $map['VersionSuggestion'];
+        }
+
+        return $model;
+    }
+}